myCode/myGAN/dga4_GAN/requirements.txt
absl-py==0.8.1
astor==0.8.0
boto==2.49.0
boto3==1.10.27
botocore==1.13.27
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
docutils==0.15.2
gast==0.2.2
gensim==3.8.1
google-pasta==0.1.7
graphviz==0.13.2
grpcio==1.24.1
h5py==2.10.0
hmmlearn==0.2.2
idna==2.8
jmespath==0.9.4
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
learn==1.0.0
Markdown==3.1.1
matplotlib==3.2.0rc1
nltk==3.4.5
numpy==1.17.2
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.0
pydot==1.4.1
pydot-ng==2.0.0
pyparsing==2.4.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.3
six==1.12.0
smart-open==1.9.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tflearn==0.3.2
torchfile==0.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
xgboost==0.90
